requirements.txt
torch==1.9.1
requests==2.25.1
opencv_python_headless==4.5.2.52
matplotlib==3.3.4
numpy==1.18.2
facenet_pytorch==2.5.2
Pillow==8.3.2
